All 0 tests removed
