+++ blinky/ble/build.gradle
@@ -12,8 +12,6 @@ android {
         minSdk android_min_api_version
         targetSdk android_api_version
         resConfigs "en"
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
 }
 
@@ -26,8 +24,4 @@ dependencies {
     implementation 'com.jakewharton.timber:timber:5.0.1'
     // Coroutines
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.6.4"
-
-    testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.3'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
 }
