+++ blinky/ble/src/main/java/no/nordicsemi/android/blinky/ble/data/ButtonCallback.kt
@@ -1,19 +1,19 @@
-package no.nordicsemi.android.blinky.transport_ble.data
+package no.nordicsemi.android.blinky.ble.data
 
 import android.bluetooth.BluetoothDevice
 import no.nordicsemi.android.ble.callback.profile.ProfileReadResponse
 import no.nordicsemi.android.ble.data.Data
 
-abstract class LedCallback: ProfileReadResponse() {
+abstract class ButtonCallback: ProfileReadResponse() {
 
     override fun onDataReceived(device: BluetoothDevice, data: Data) {
         if (data.size() == 1) {
-            val ledState = data.getIntValue(Data.FORMAT_UINT8, 0) == 0x01
-            onLedStateChanged(device, ledState)
+            val buttonState = data.getIntValue(Data.FORMAT_UINT8, 0) == 0x01
+            onButtonStateChanged(device, buttonState)
         } else {
             onInvalidDataReceived(device, data)
         }
     }
 
-    abstract fun onLedStateChanged(device: BluetoothDevice, state: Boolean)
+    abstract fun onButtonStateChanged(device: BluetoothDevice, state: Boolean)
 }
