+++ blinky/ui/build.gradle
@@ -13,8 +13,6 @@ android {
         minSdk android_min_api_version
         targetSdk android_api_version
         resConfigs "en"
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
 
     /**
@@ -52,10 +50,6 @@ dependencies {
     implementation "androidx.activity:activity-compose:1.6.0"
     // Extended icons
     implementation "androidx.compose.material:material-icons-extended:$compose_version"
-
-    testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.3'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
 
     // Hilt and Dagger
     implementation 'com.google.dagger:hilt-android:2.43.2'
