+++ scanner/build.gradle
@@ -12,8 +12,6 @@ android {
         minSdk android_min_api_version
         targetSdk android_api_version
         resConfigs "en"
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
 
     /**
@@ -45,8 +43,4 @@ dependencies {
     implementation "androidx.compose.material3:material3:1.0.0-beta03"
     // Integration with activities
     implementation "androidx.activity:activity-compose:1.6.0"
-
-    testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.3'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
 }
